move shared analyze_result functions to main library
