+++ cpp/models/secir_vaccine/analyze_result.h
@@ -1,7 +1,7 @@
 /* 
 * Copyright (C) 2020-2021 German Aerospace Center (DLR-SC)
 *
-* Authors: Daniel Abele
+* Authors: Wadim Koslow, Daniel Abele
 *
 * Contact: Martin J. Kuehn <Martin.Kuehn@DLR.de>
 *
@@ -17,12 +17,11 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-#ifndef EPI_SECIR_ANALYZE_RESULT_H
-#define EPI_SECIR_ANALYZE_RESULT_H
+#ifndef MEMILIO_SECIR_VACC_ANALYZE_RESULT_H
+#define MEMILIO_SECIR_VACC_ANALYZE_RESULT_H
 
-#include "memilio/utils/time_series.h"
 #include "secir_vaccine/secir.h"
-#include "memilio/mobility/mobility.h"
+#include "memilio/data/analyze_result.h"
 
 #include <functional>
 #include <vector>
@@ -31,91 +30,12 @@ namespace mio
 {
 namespace vaccinated
 {
-
     /**
- * interpolate time series with evenly spaced, integer time points.
- * time points [t0, t1, t2, ..., tmax] interpolated as [floor(t0), floor(t0) + 1,...,ceil(tmax)].
- * values at new time points are linearly interpolated from their immediate neighbors from the old time points.
- * @param simulation_result time series to interpolate
- * @return interpolated time series
- */
-    TimeSeries<double> interpolate_simulation_result(const TimeSeries<double>& simulation_result);
-
-    /**
- * helper template, type returned by overload interpolate_simulation_result(T t)
- */
-    template <class T>
-    using InterpolateResultT = std::decay_t<decltype(interpolate_simulation_result(std::declval<T>()))>;
-
-    /**
- * @brief Interpolates results of all runs with evenly spaced, integer time points.
- * @see interpolate_simulation_result
- * @param ensemble_result result of multiple simulations (single TimeSeries or Graph)
- * @return interpolated time series, one (or as many as nodes in the graph) per result in the ensemble
- */
-    template <class T>
-    std::vector<InterpolateResultT<T>> interpolate_ensemble_results(const std::vector<T>& ensemble_results)
-    {
-        std::vector<InterpolateResultT<T>> interpolated;
-        interpolated.reserve(ensemble_results.size());
-        std::transform(ensemble_results.begin(), ensemble_results.end(), std::back_inserter(interpolated),
-                       [](auto& run) {
-                           return interpolate_simulation_result(run);
-                       });
-        return interpolated;
-    }
-
-    std::vector<std::vector<TimeSeries<double>>>
-    sum_nodes(const std::vector<std::vector<TimeSeries<double>>>& ensemble_result);
-
-    /**
- * @brief computes mean of each compartment, node, and time point over all runs
- * input must be uniform as returned by interpolated_ensemble_result:
- * same number of nodes, same time points and elements.
- * @see interpolated_ensemble_result
- * @param ensemble_results uniform results of multiple simulation runs
- * @return mean of the results over all runs
- */
-    std::vector<TimeSeries<double>> ensemble_mean(const std::vector<std::vector<TimeSeries<double>>>& ensemble_results);
-
-    /**
- * @brief computes the p percentile of the result for each compartment, node, and time point.
- * Produces for each compartment the value that that is bigger than approximately
- * a p-th share of the values of this compartment over all runs.
- * input must be uniform as returned by interpolated_ensemble_result:
- * same number of nodes, same time points and elements.
- * @see interpolated_ensemble_result
- * @param ensemble_result uniform results of multiple simulation runs
- * @param p percentile value in open interval (0, 1)
- * @return p percentile of the results over all runs
- */
-    std::vector<TimeSeries<double>>
-    ensemble_percentile(const std::vector<std::vector<TimeSeries<double>>>& ensemble_result, double p);
-    /**
- * interpolate time series with evenly spaced, integer time points for each node.
- * @see interpolate_simulation_result
- * @param graph_result graph of simulations whose results will be interpolated
- * @return one interpolated time series per node
- */
-    template <class Simulation>
-    std::vector<TimeSeries<double>>
-    interpolate_simulation_result(const Graph<SimulationNode<Simulation>, MigrationEdge>& graph_result)
-    {
-        std::vector<TimeSeries<double>> interpolated;
-        interpolated.reserve(graph_result.nodes().size());
-        std::transform(graph_result.nodes().begin(), graph_result.nodes().end(), std::back_inserter(interpolated),
-                       [](auto& n) {
-                           return interpolate_simulation_result(n.property.get_result());
-                       });
-        return interpolated;
-    }
-
-    /**
- * @brief computes the p percentile of the parameters for each node.
- * @param ensemble_result graph of multiple simulation runs
- * @param p percentile value in open interval (0, 1)
- * @return p percentile of the parameters over all runs
- */
+     * @brief computes the p percentile of the parameters for each node.
+     * @param ensemble_result graph of multiple simulation runs
+     * @param p percentile value in open interval (0, 1)
+     * @return p percentile of the parameters over all runs
+     */
     template <class Model>
     std::vector<Model> ensemble_params_percentile(const std::vector<std::vector<Model>>& ensemble_params, double p)
     {
@@ -223,30 +143,7 @@ namespace vaccinated
         return percentile;
     }
 
-    /**
- * Compute the distance between two SECIR simulation results.
- * The distance is the 2-norm of the element-wise difference of the two results.
- * The two results (e.g. output of interpolate_simulation_result) must have the same dimensions and number of time points.
- * @param result1 first result.
- * @param result2 second result.
- * @return Computed distance between result1 and result2.
- */
-    double result_distance_2norm(const std::vector<mio::TimeSeries<double>>& result1,
-                                 const std::vector<mio::TimeSeries<double>>& result2);
-
-    /**
- * Compute the distance between two SECIR simulation results in one compartment.
- * The distance is the 2-norm of the element-wise difference of the two results in the specified compartment.
- * The two results (e.g. output of interpolate_simulation_result) must have the same dimensions and number of time points.
- * @param result1 first result.
- * @param result2 second result.
- * @param compartment the compartment to compare.
- * @return Computed distance between result1 and result2.
- */
-    double result_distance_2norm(const std::vector<mio::TimeSeries<double>>& result1,
-                                 const std::vector<mio::TimeSeries<double>>& result2, InfectionState compartment);
-
 } // namespace vaccinated
 } // namespace mio
 
-#endif //EPI_SECIR_ANALYZE_RESULT_H
+#endif //MEMILIO_SECIR_VACC_ANALYZE_RESULT_H
